Skrevet tester for medlemskap
